requirements.txt
uvicorn
fastapi
python-multipart
email-validator
passlib[bcrypt]
python-jose
python-jose[cryptography]
psycopg2-binary
alembic
sqlalchemy
emails

sendgrid
python-dotenv

pytest
pytest-cov